[docs] Update readme for golang server remake
