+++ old-launcher/requirements.txt
@@ -1,0 +1,2 @@
+watchdog
+psutil
